Implemented hand movement detection
This is used for activating and deactivating the raycasting (pointing detection) during response time according to the hand movement. The pointing detection will now only work if the hand does not move (movement is under the threshold). This whole feature can be activated or deactivated.
Additionally the pointing detection is now only switched on during reaction phase in general.